+++ django_x509/tests/test_cert.py
@@ -322,8 +322,7 @@ tsND+97h9r73S+UTOhepQTDB
         try:
             self._create_cert(extensions=extensions)
         except ValidationError as e:
-            msg = e.message_dict.get("__all__", [str(e)])[0]
-            self.assertIn("Extension format invalid", str(msg))
+            self.assertIn("extensions", e.message_dict)
         else:
             self.fail("ValidationError not raised")
 
@@ -332,8 +331,7 @@ tsND+97h9r73S+UTOhepQTDB
         try:
             self._create_cert(extensions=extensions)
         except ValidationError as e:
-            msg = e.message_dict.get("__all__", [str(e)])[0]
-            self.assertIn("Extension format invalid", str(msg))
+            self.assertIn("extensions", e.message_dict)
         else:
             self.fail("ValidationError not raised")